# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eltuvchi API",
        "description": "Eltuvchi platformasi uchun API hujjatlari",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/super-admin/market": {
            "get": {
                "tags": [
                    "Admin Market"
                ],
                "summary": "Get list of products or discounts",
                "operationId": "b7219518189715726d0e87a8ebe8319f",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of items to fetch (product|discount)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "product",
                                "discount"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of items"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin Market"
                ],
                "summary": "Create product or discount",
                "operationId": "0ade1e3caa3b95536eb12cd1b34e920d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "product",
                                            "discount"
                                        ]
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "icon_type": {
                                        "type": "string"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "value": {
                                        "type": "integer"
                                    },
                                    "percent": {
                                        "type": "integer"
                                    },
                                    "icon": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    }
                }
            }
        },
        "/api/super-admin/market/{id}": {
            "put": {
                "tags": [
                    "Admin Market"
                ],
                "summary": "Update product or discount (only title, description, status)",
                "operationId": "5b80358e95dcb3db7e01b3fe30a18a4f",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of item (product|discount)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "product",
                                "discount"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    }
                }
            }
        },
        "/api/admin/orders": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Get all orders",
                "operationId": "3bd2889661cdab2ffdd9a3c26aee96c6",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter orders by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "accepted",
                                "completed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdminOrder"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/orders/{id}": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Get single order by ID",
                "operationId": "266192160779cd669287f02bb5b3b733",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminOrder"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Update order by ID",
                "operationId": "b68bca16067b7073a1ae85e37d95b66b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "completed"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Admin updated order"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminOrder"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Cancel order by ID",
                "operationId": "1a202ce7b504d56699c0adc8305eb049",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order cancelled successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AdminOrder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AdminClient": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "integer"
                    },
                    "points": {
                        "type": "integer"
                    },
                    "settings": {
                        "type": "object"
                    },
                    "user": {
                        "type": "object"
                    },
                    "discounts": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "orders_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "AdminDriver": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "integer"
                    },
                    "points": {
                        "type": "integer"
                    },
                    "details": {
                        "type": "object"
                    },
                    "settings": {
                        "type": "object"
                    },
                    "user": {
                        "type": "object"
                    },
                    "taxopark": {
                        "type": "object"
                    },
                    "orders_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "AdminOrder": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "scheduled_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-20 14:00:00"
                    },
                    "passengers": {
                        "type": "integer",
                        "example": 3
                    },
                    "price_summ": {
                        "type": "number",
                        "format": "float",
                        "example": 150000
                    },
                    "client_deposit": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "price_order": {
                        "type": "number",
                        "format": "float",
                        "example": 200000
                    },
                    "phone": {
                        "type": "string",
                        "example": "+998901234567"
                    },
                    "optional_phone": {
                        "type": "string",
                        "example": "+998931234567",
                        "nullable": true
                    },
                    "note": {
                        "type": "string",
                        "example": "Urgent order",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-15T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T10:00:00Z"
                    },
                    "client": {
                        "$ref": "#/components/schemas/AdminClient"
                    },
                    "driver": {
                        "$ref": "#/components/schemas/AdminDriver"
                    },
                    "route": {
                        "$ref": "#/components/schemas/Route"
                    },
                    "histories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderHistory"
                        }
                    },
                    "review": {
                        "$ref": "#/components/schemas/Review"
                    }
                },
                "type": "object"
            },
...
            "OrderHistory": {
                "title": "Order History Resource",
                "description": "Order status changes history",
                "properties": {
                    "id": {
                        "description": "History record ID",
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "Order status at this history record",
                        "type": "string",
                        "example": "pending"
                    },
                    "description": {
                        "description": "Reason or note for the status change",
                        "type": "string",
                        "example": "Order created by client"
                    },
                    "changed_by_type": {
                        "description": "Who changed the status (client, driver, admin)",
                        "type": "string",
                        "example": "client"
                    },
                    "changed_by_id": {
                        "description": "ID of the user who changed the status",
                        "type": "integer",
                        "example": 42
                    },
                    "created_at": {
                        "description": "When this history record was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-14 12:30:45"
                    }
                },
                "type": "object"
            },
...
            "Region": {
                "title": "Region Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Tashkent"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-14T10:30:00Z"
                    }
                },
                "type": "object"
            },
            "Review": {
                "title": "Review Resource",
                "description": "Review details for orders or items",
                "properties": {
                    "rating": {
                        "description": "Rating given, e.g., from 1 to 5",
                        "type": "integer",
                        "example": 5
                    },
                    "comment": {
                        "description": "Optional comment for the review",
                        "type": "string",
                        "example": "Great service!",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Route": {
                "title": "Route Resource",
                "description": "Route details including taxopark, pricing and status",
                "properties": {
                    "id": {
                        "description": "Route ID",
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "$ref": "#/components/schemas/Taxopark"
                    },
                    "to": {
                        "$ref": "#/components/schemas/Taxopark"
                    },
                    "status": {
                        "description": "Route status",
                        "type": "string",
                        "example": "active"
                    },
                    "deposit_client": {
                        "description": "Client deposit amount for this route",
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "distance_km": {
                        "description": "Distance in kilometers",
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "price_in": {
                        "description": "Price for the route",
                        "type": "number",
                        "format": "float",
                        "example": 120000
                    },
                    "fee_per_client": {
                        "description": "Fee per client",
                        "type": "number",
                        "format": "float",
                        "example": 10000
                    }
                },
                "type": "object"
            },
            "Taxopark": {
                "title": "TaxoPark",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "region": {
                        "$ref": "#/components/schemas/Region"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "name": {
                        "type": "string",
                        "example": "Toshkent Taxi"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum authentication using Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}